requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Flask-Migrate
Flask-WTF
python-dotenv
werkzeug
pymysql==1.1.1
pytz